+++ src/App.tsx
@@ -4,11 +4,10 @@ import { getConfig } from '@/config';
 import { setupAxiosInterceptors } from '@api/auth';
 import { ProtectedRoute, Signin, Signup } from '@auth/index';
 import { AuthProvider } from '@common/contexts/auth';
-import { NotesProvider } from '@notes/context';
 import Notes from '@notes/Notes';
 import Planner from '@planner/Planner';
 import { Layout } from '@sections/Layout';
-import { Settings, SettingsProfile, SettingsPlanner, SettingsNotes } from '@settings/index';
+import { Settings, SettingsGeneral, SettingsPlanner, SettingsNotes } from '@settings/index';
 
 // Setup axios interceptors for token inclusion before mounting the App
 setupAxiosInterceptors();
@@ -34,17 +33,10 @@ function App() {
               <Route path="planner" element={<Planner />} />
               <Route path="notes" element={<Notes />} />
               <Route path="settings" element={<Settings />}>
-                <Route index element={<Navigate to="/settings/profile/" replace />} />
-                <Route path="profile" element={<SettingsProfile />} />
+                <Route index element={<Navigate to="/settings/general/" replace />} />
+                <Route path="general" element={<SettingsGeneral />} />
                 <Route path="planner" element={<SettingsPlanner />} />
-                <Route
-                  path="notes"
-                  element={
-                    <NotesProvider>
-                      <SettingsNotes />
-                    </NotesProvider>
-                  }
-                />
+                <Route path="notes" element={<SettingsNotes />} />
               </Route>
             </Route>
           </Route>