+++ src/sections/settings/index.ts
@@ -1,4 +1,4 @@
 export * from './Settings';
-export * from './components/SettingsProfile/SettingsProfile';
+export * from './components/SettingsGeneral/SettingsGeneral';
 export * from './components/SettingsPlanner/SettingsPlanner';
 export * from './components/SettingsNotes/SettingsNotes';